refactor: rename `named_prelude_import` to `named_prelude_imports`
Pluralize the rule name to match `wildcard_imports` and the plural form
many Clippy lints use, so `#[allow(perfectionist::named_prelude_imports)]`
reads naturally. Renames the module and its files, the lint ident, the
config key, the ui/ui-toml fixtures and tests, the generated rules
markdown, and the planning file, and updates the `wildcard_imports`
cross-reference.
