+++ rules/named_prelude_imports.md
@@ -1,9 +1,9 @@
 <!-- Generated by `gen-docs write-md` from `src/rules/`. Do not edit by hand; run `just gen-rules-md` after changing a rule source. -->
 
-# `perfectionist::named_prelude_import`
+# `perfectionist::named_prelude_imports`
 
 **Default state:** `active`  
-**Source:** [`src/rules/named_prelude_import.rs`](../src/rules/named_prelude_import.rs)
+**Source:** [`src/rules/named_prelude_imports.rs`](../src/rules/named_prelude_imports.rs)
 
 > named item cherry-picked from a prelude module instead of glob-imported
 
@@ -60,7 +60,7 @@ use diesel::prelude::*;
 
 ## Configuration
 
-Configure via `dylint.toml` under `["perfectionist::named_prelude_import"]`. Every field is optional; the per-field prose below states the default.
+Configure via `dylint.toml` under `["perfectionist::named_prelude_imports"]`. Every field is optional; the per-field prose below states the default.
 
 ### `prelude_segment_names`: `[string]` (optional)
