+++ src/lib.rs
@@ -59,7 +59,7 @@ pub fn register_lints(session: &Session, lint_store: &mut LintStore) {
         lint_silence_reason
         macro_argument_binding
         macro_trailing_comma
-        named_prelude_import
+        named_prelude_imports
         non_exhaustive_error
         prefer_derive_more_over_thiserror
         prefer_expect_over_allow
