+++ ui/named_prelude_imports.stderr
@@ -1,19 +1,19 @@
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:22:9
+  --> $DIR/named_prelude_imports.rs:22:9
    |
 LL |     use crate::prelude::A;
    |         ^^^^^^^^^^^^^^^^^ help: import the item from its canonical module: `crate::thing::A`
    |
-   = note: `#[warn(perfectionist::named_prelude_import)]` on by default
+   = note: `#[warn(perfectionist::named_prelude_imports)]` on by default
 
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:27:9
+  --> $DIR/named_prelude_imports.rs:27:9
    |
 LL |     use crate::prelude::B as Renamed;
    |         ^^^^^^^^^^^^^^^^^ help: import the item from its canonical module: `crate::thing::B`
 
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:32:26
+  --> $DIR/named_prelude_imports.rs:32:26
    |
 LL |     use crate::prelude::{A, helper};
    |                          ^
@@ -21,7 +21,7 @@ LL |     use crate::prelude::{A, helper};
    = help: import this item from its canonical module, or glob-import the prelude with `use ...::prelude::*;`
 
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:32:29
+  --> $DIR/named_prelude_imports.rs:32:29
    |
 LL |     use crate::prelude::{A, helper};
    |                             ^^^^^^
@@ -29,7 +29,7 @@ LL |     use crate::prelude::{A, helper};
    = help: import this item from its canonical module, or glob-import the prelude with `use ...::prelude::*;`
 
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:65:9
+  --> $DIR/named_prelude_imports.rs:65:9
    |
 LL |     use crate::multi::prelude::Dual;
    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -37,7 +37,7 @@ LL |     use crate::multi::prelude::Dual;
    = help: import this item from its canonical module, or glob-import the prelude with `use ...::prelude::*;`
 
 warning: named item cherry-picked from a prelude module
-  --> $DIR/named_prelude_import.rs:81:9
+  --> $DIR/named_prelude_imports.rs:81:9
    |
 LL |     use crate::kw_prelude::prelude::Thing;
    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: import the item from its canonical module: `crate::r#type::Thing`
